% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckacorrdoe.R
\name{ckacorrdoe}
\alias{ckacorrdoe}
\title{To Identify Outliers in Design of Experiments with Autocorrelated Errors under lag 1}
\usage{
ckacorrdoe(trt, Rep, resp, noutli)
}
\arguments{
\item{trt}{Numeric or complex vector containing the treatment levels}

\item{Rep}{Numeric or complex vector containing the Replication}

\item{resp}{Numeric or complex vector containing response variable}

\item{noutli}{a  number indicating the number of outliers}
}
\value{
The output contains Shapiro-Wilk Normality test and Bartlett test for
Residuals of the of the model and Cook's distance for each treatment or a
combination of treatments in Design of Experiments with auto-correlated errors under lag 1
}
\description{
To Identify Outliers in Design of Experiments with Autocorrelated Errors under lag 1
}
\examples{
data(ex3)
attach(ex3)
ckacorrdoe(ex3$trt,ex3$rep,ex3$`sim 1`, 1)

}
