% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckgendoe.R
\name{ckgdoe}
\alias{ckgdoe}
\title{To Identify Outliers in Design of Experiments}
\usage{
ckgdoe(trt, Rep, resp, noutli)
}
\arguments{
\item{trt}{Numeric or complex vector containing the treatment levels}

\item{Rep}{Numeric or complex vector containing the Replication}

\item{resp}{Numeric or complex vector containing response variable}

\item{noutli}{a number indicating the number of outliers}

\item{threshold}{}
}
\value{
The output contains Shapiro-Wilk Normality test and Bartlett test for
Residuals of the of the model and Cook's distance for each treatment or a
combination of treatments
Title
}
\description{
To Identify Outliers in Design of Experiments
}
