% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckgendoe.R
\name{ckgendoe}
\alias{ckgendoe}
\title{To Identify Outliers in Design of Experiments}
\usage{
ckgendoe(trt, Rep, resp, noutli)
}
\arguments{
\item{trt}{Numeric or complex vector containing the treatment levels}

\item{Rep}{Numeric or complex vector containing the Replication}

\item{resp}{Numeric or complex vector containing response variable}

\item{noutli}{a number indicating the number of outliers}
}
\value{
The output contains Shapiro-Wilk Normality test and Bartlett test for
Residuals of the of the model and Cook's distance for each treatment or a
combination of treatments
}
\description{
To Identify Outliers in Design of Experiments
}
\examples{
# example code
data(ex1)
ckgendoe(ex1$trt,exq$rep,exq$yld,1)
}
