% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmultdoe.R
\name{ckmultdoe}
\alias{ckmultdoe}
\title{To Identify Outliers in Multi - Response Experiments}
\usage{
ckmultdoe(trt, Rep, resps, noutli)
}
\arguments{
\item{trt}{Numeric or complex vector containing the treatment levels}

\item{Rep}{Numeric or complex vector containing the Replication}

\item{resps}{data frame containing response variables}

\item{noutli}{a  number indicating the number of outliers}
}
\value{
The output contains Shapiro-Wilk Normality test and Bartlett test for
Residuals of the of the model and Cook's distance for each treatment or a
combination of treatments in Multi - Response Experiments
}
\description{
To Identify Outliers in Multi - Response Experiments
}
\examples{
data(ex2)
attach(ex2)
ckmultdoe(ex2$trt,ex2$rep,ex2[,3:4],1)

}
