% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ex1.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{Sample data for analysis}
\format{
A \code{tibble} with 3 column, which are:
\describe{
\item{Rep}{Numeric or complex vector of Replication}
\item{trt}{Numeric or complex vector of treatments}
\item{yld}{Numeric vector of responses}
}

A \code{tibble} with 3 column, which are:
\describe{
\item{Rep}{Numeric or complex vector of Replication}
\item{trt}{Numeric or complex vector of treatments}
\item{yld}{Numeric vector of responses}
}
}
\source{
{created ex1 to serve as an example}

{created ex1 to serve as an example}
}
\usage{
ex1

ex1
}
\description{
A simple \code{tibble} that is for simple testing and running
of examples
The data is taken from A Textbook of Agricultural Statistics by R Rangaswamy

A simple \code{tibble} that is for simple testing and running
of examples
The data is taken from A Textbook of Agricultural Statistics by R Rangaswamy
}
\examples{
data(ex1)
data(ex1)
}
\keyword{datasets}
