% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/ex2.R
\docType{data}
\name{ex2}
\alias{ex2}
\title{Sample data for analysis for multi-response experiments}
\format{
A \code{tibble} with 4 column, which are:
\describe{
\item{Rep}{Numeric or complex vector of Replication}
\item{trt}{Numeric or complex vector of treatments}
\item{yld1}{Numeric vector of response 1}
\item{yld2}{Numeric vector of response 2}
}

A \code{tibble} with 4 column, which are:
\describe{
\item{rep}{Numeric or complex vector of Replication}
\item{trt}{Numeric or complex vector of treatments}
\item{yld1}{Numeric vector of response 1}
\item{yld2}{Numeric vector of response 2}
}
}
\source{
{created ex1 to serve as an example}

{created ex2 to serve as an example}
}
\usage{
ex2

ex2
}
\description{
A simple \code{tibble} used as example for multi - response
experiments

A simple \code{tibble} used as example for multi - response
experiments
}
\examples{
data(ex2)
data(ex2)
}
\keyword{datasets}
