% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex3.R
\docType{data}
\name{ex3}
\alias{ex3}
\title{Sample data for analysis for Design of Experiments with Auto-correlated Errors}
\format{
A \code{tibble} with 4 column, which are:
\describe{
\item{rep}{Numeric or complex vector of Replication}
\item{trt}{Numeric or complex vector of treatments}
\item{`sim 1`}{Numeric vector serves as response 1}
\item{`sim 2`}{Numeric vector serves as response 2}
}
}
\source{
{created ex3 to serve as an example}
}
\usage{
ex3
}
\description{
A simple \code{tibble} used as example for Design of Experiments
with Auto-correlated Errors

This data is simulated for example purpose
}
\examples{
data(ex3)
}
\keyword{datasets}
