% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckgendoe.R
\name{ckgendoe}
\alias{ckgendoe}
\title{To Identify Outliers in Design of Experiments}
\usage{
ckgendoe(trt, Rep, resp, noutli)
}
\arguments{
\item{trt}{Numeric or complex vector containing the treatment levels}

\item{Rep}{Numeric or complex vector containing the Replication}

\item{resp}{Numeric or complex vector containing response variable}

\item{noutli}{a number indicating the number of outliers}
}
\value{
The output contains Shapiro-Wilk Normality test and Bartlett test for
Residuals of the of the model and Cook's distance for each treatment or a
combination of treatments

 @examples
 data(ex1)
 To identify single outlier in block designs
 ckgendoe(ex1$trt,ex1$rep,ex1$yld,1)

 @export
}
\description{
To Identify Outliers in Design of Experiments
}
